requirements.txt
asgiref==3.2.3
Django==3.0.4
django-js-asset==1.2.2
django-reversion==3.0.7
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.1
